editor/libeditor/tests/test_execCommandPaste_noTarget.html

<!DOCTYPE HTML>
<html>
<head>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script>

  add_task(async function() {
    let seenPaste = false;
    let seenCopy = false;
    document.addEventListener("copy", function oncpy(e) {
      document.removeEventListener("copy", oncpy);
      e.clipboardData.setData("text/plain", "my text");
      e.preventDefault();
      seenCopy = true;
    });
    document.addEventListener("paste", function onpst(e) {
      document.removeEventListener("paste", onpst);
      is(e.clipboardData.getData("text/plain"), "my text",
         "The correct text was read from the clipboard");
      e.preventDefault();
      seenPaste = true;
    });

    ok(SpecialPowers.wrap(document).execCommand("copy"),
       "Call should succeed");
    ok(seenCopy, "Successfully copied the text to the clipboard");
    ok(SpecialPowers.wrap(document).execCommand("paste"),
       "Call should succeed");
    ok(seenPaste, "Successfully read text from the clipboard");

    // Check that reading text from the clipboard in non-privileged contexts
    // still doesn't work.
    function onpstfail(e) {
      ok(false, "Should not see paste event triggered by non-privileged call");
    }
    document.addEventListener("paste", onpstfail);
    ok(!document.execCommand("paste"), "Call should fail");
    document.removeEventListener("paste", onpstfail);
  });

</script>
</body>
</html>
